Move migrate-consolidations to use-case dataflows
Jira ticket: CAMS-421
