+++ backend/function-apps/dataflows/migration/activity/flattenBoundingArrays.test.ts
@@ -2,7 +2,7 @@ import { InvocationContext } from '@azure/functions';
 import {
   AcmsBounds,
   AcmsPredicate,
-} from '../../../../lib/use-cases/acms-orders/migrate-consolidations';
+} from '../../../../lib/use-cases/dataflows/migrate-consolidations';
 import module from './flattenBoundingArrays';
 
 describe('Flatten bounding arrays activity', () => {
