+++ backend/function-apps/dataflows/migration/activity/queueMigrateConsolidation.ts
@@ -2,7 +2,7 @@ import { InvocationContext, output } from '@azure/functions';
 import {
   AcmsEtlQueueItem,
   AcmsPredicate,
-} from '../../../../lib/use-cases/acms-orders/migrate-consolidations';
+} from '../../../../lib/use-cases/dataflows/migrate-consolidations';
 import ContextCreator from '../../../azure/application-context-creator';
 import AcmsOrdersController from '../../../../lib/controllers/acms-orders/acms-orders.controller';
 import { getCamsError } from '../../../../lib/common-errors/error-utilities';
