+++ backend/function-apps/dataflows/migration/queueTrigger/migrateConsolidation.ts
@@ -3,7 +3,7 @@ import ContextCreator from '../../../azure/application-context-creator';
 import AcmsOrdersController from '../../../../lib/controllers/acms-orders/acms-orders.controller';
 import { getCamsError } from '../../../../lib/common-errors/error-utilities';
 import { CamsError } from '../../../../lib/common-errors/cams-error';
-import { isAcmsEtlQueueItem } from '../../../../lib/use-cases/acms-orders/migrate-consolidations';
+import { isAcmsEtlQueueItem } from '../../../../lib/use-cases/dataflows/migrate-consolidations';
 
 const MODULE_NAME = 'IMPORT_ACTION_MIGRATE_CONSOLIDATION';
