+++ backend/lib/adapters/gateways/acms/acms.gateway.test.ts
@@ -5,7 +5,7 @@ import {
   AcmsConsolidation,
   AcmsConsolidationChildCase,
   AcmsPredicate,
-} from '../../../use-cases/acms-orders/migrate-consolidations';
+} from '../../../use-cases/dataflows/migrate-consolidations';
 import { UnknownError } from '../../../common-errors/unknown-error';
 
 describe('ACMS gateway tests', () => {
